+++ server/controllers/eventController.js
@@ -1,0 +1,75 @@
+const Table = require('../models/tableModels');
+const existed = {status: 'already exists'};
+
+
+
+const createEvent = (req, res) => {
+  Table.Event.findOrCreate({
+    where: {
+      eventName: req.body.eventName,
+      password: req.body.password,
+      latitude: req.body.latitude,
+      longitude: req.body.longitude,
+      userId: req.params.userId
+    }
+  })
+  .spread((response, isCreated) => {
+    if (isCreated) {
+      res.status(201).send(response);
+    } else {
+      res.send(existed);
+    }
+  })
+  .catch((error) => {
+    res.send(error);
+  });
+};
+
+const fetchUserEvents = (req, res) => {
+  Table.User_Event.findAll({
+    where: {
+      userId: req.params.userId
+    },
+    include: [
+      {
+        model: Table.Event
+      }
+    ]
+  })
+  .then((response) => {
+    res.status(200).send(response);
+  })
+  .catch((error) => {
+    res.send(error);
+  });
+};
+
+
+
+
+const joinEvent = (req, res) => {
+  Table.User_Event.findOrCreate({
+    where: {
+      userId: req.params.userId,
+      eventId: req.body.eventId
+    }
+  })
+  .spread((response, isCreated) => {
+    if (isCreated) {
+      res.status(201).send(response);
+    } else {
+      res.send(existed);
+    }
+  })
+  .catch((error) => {
+    res.send(error);
+  });
+};
+
+
+module.exports = {
+
+  createEvent: createEvent,
+  fetchUserEvents: fetchUserEvents,
+  joinEvent: joinEvent
+}
